+++ views/layout.hbs
@@ -186,11 +186,11 @@
 					<img src="images/icons/icon-close2.png" alt="CLOSE">
 				</button>
 
-				<form action="/" metod = "GET" class="wrap-search-header flex-w p-l-15">
+				<form action="/search-title" metod = "GET" class="wrap-search-header flex-w p-l-15">
 					<button class="flex-c-m trans-04">
 						<i class="zmdi zmdi-search"></i>
 					</button>
-					<input class="plh3" type="text" name="search" placeholder="Nhập tên sách cần tìm">
+					<input class="plh3" type="text" name="search_title" placeholder="Nhập tên sách cần tìm">
 				</form>
 			</div>
 		</div>
