+++ views/popup-page.hbs
@@ -26,7 +26,7 @@
         </h4>
 
         <span class="mtext-106 cl2">
-            {{ this.price }} ₫
+            {{ this.price }}₫
         </span>
 
         <p class="stext-102 cl3 p-t-23">
