+++ views/product-detail.hbs
@@ -1,12 +1,12 @@
 
 <div class="container">
 	<div class="bread-crumb flex-w p-l-25 p-r-15 p-t-30 p-lr-0-lg">
-		<a href="index" class="stext-109 cl8 hov-cl1 trans-04">
+		<a href="/" class="stext-109 cl8 hov-cl1 trans-04">
 			Trang chủ
 			<i class="fa fa-angle-right m-l-9 m-r-10" aria-hidden="true"></i>
 		</a>
 
-		<a href="" class="stext-109 cl8 hov-cl1 trans-04">
+		<a href="#" class="stext-109 cl8 hov-cl1 trans-04">
 			{{publisher}}
 			<i class="fa fa-angle-right m-l-9 m-r-10" aria-hidden="true"></i>
 		</a>
